remove deprecated members
