+++ packages/jaspr/lib/src/foundation/styles/properties/box.dart
@@ -29,8 +29,6 @@ enum Display {
 abstract class Border {
   const factory Border({BorderStyle? style, Color? color, Unit? width}) = __Border;
 
-  @Deprecated('Use Border() instead.')
-  const factory Border.all(BorderSide side) = _AllBorder;
   const factory Border.only({BorderSide? left, BorderSide? top, BorderSide? right, BorderSide? bottom}) = _OnlyBorder;
   const factory Border.symmetric({BorderSide? vertical, BorderSide? horizontal}) = _SymmetricBorder;
